# Added GROUP-BY-LICENSE argument + made function 1000 times faster